Android: moves ui/ to use long for JNI.
BUG=

Review URL: https://codereview.chromium.org/61733010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@234312 0039d316-1c4b-4281-b951-d872f2087c98

+++ content/public/android/java/src/org/chromium/content/browser/ContentViewCore.java
@@ -753,9 +753,9 @@ public class ContentViewCore
             }
         };
 
-        int windowNativePointer = windowAndroid != null ? windowAndroid.getNativePointer() : 0;
+        long windowNativePointer = windowAndroid != null ? windowAndroid.getNativePointer() : 0;
 
-        int viewAndroidNativePointer = 0;
+        long viewAndroidNativePointer = 0;
         if (windowNativePointer != 0) {
             mViewAndroid = new ViewAndroid(windowAndroid, getViewAndroidDelegate());
             viewAndroidNativePointer = mViewAndroid.getNativePointer();
@@ -3209,7 +3209,7 @@ public class ContentViewCore
     }
 
     private native int nativeInit(boolean hardwareAccelerated, int webContentsPtr,
-            int viewAndroidPtr, int windowAndroidPtr);
+            long viewAndroidPtr, long windowAndroidPtr);
 
     @CalledByNative
     private ContentVideoViewClient getContentVideoViewClient() {
