+++ ui/android/java/src/org/chromium/ui/gfx/SurfaceTexturePlatformWrapper.java
@@ -30,7 +30,7 @@ class SurfaceTexturePlatformWrapper {
 
     @CalledByNative
     private static void setFrameAvailableCallback(SurfaceTexture surfaceTexture,
-            int nativeSurfaceTextureListener) {
+            long nativeSurfaceTextureListener) {
        surfaceTexture.setOnFrameAvailableListener(
                new SurfaceTextureListener(nativeSurfaceTextureListener));
     }
